refactor iast and exploit prevention
