+++ ddtrace/appsec/_iast/taint_sinks/__init__.py
@@ -1,8 +1,6 @@
 from .ast_taint import ast_function
-from .path_traversal import open_path_traversal
 
 
 __all__ = [
-    "open_path_traversal",
     "ast_function",
 ]
