+++ ddtrace/appsec/_iast/taint_sinks/path_traversal.py
@@ -4,10 +4,7 @@ from ddtrace.internal.logger import get_logger
 
 from ..._constants import IAST_SPAN_TAGS
 from .. import oce
-from .._metrics import _set_metric_iast_instrumented_sink
 from .._metrics import increment_iast_span_metric
-from .._patch import set_and_check_module_is_patched
-from .._patch import set_module_unpatched
 from ..constants import EVIDENCE_PATH_TRAVERSAL
 from ..constants import VULN_PATH_TRAVERSAL
 from ..processor import AppSecIastSpanProcessor
@@ -36,19 +33,6 @@ def get_version():
     return ""
 
 
-def unpatch_iast():
-    # type: () -> None
-    set_module_unpatched("builtins", default_attr="_datadog_path_traversal_patch")
-
-
-def patch():
-    # type: () -> None
-    """Wrap functions which interact with file system."""
-    if not set_and_check_module_is_patched("builtins", default_attr="_datadog_path_traversal_patch"):
-        return
-    _set_metric_iast_instrumented_sink(VULN_PATH_TRAVERSAL)
-
-
 def check_and_report_path_traversal(*args: Any, **kwargs: Any) -> None:
     if AppSecIastSpanProcessor.is_span_analyzed() and PathTraversal.has_quota():
         try:
@@ -63,8 +47,3 @@ def check_and_report_path_traversal(*args: Any, **kwargs: Any) -> None:
             log.debug("Unexpected exception while reporting vulnerability", exc_info=True)
     else:
         log.debug("IAST: no vulnerability quota to analyze more sink points")
-
-
-def open_path_traversal(*args, **kwargs):
-    check_and_report_path_traversal(*args, **kwargs)
-    return open(*args, **kwargs)
